refactor: consolidate duplicate base readers and clean up architecture
- Enhanced core/base_reader.py with constructor and context manager support
- Moved Bruker-specific exceptions from readers/bruker/core/ to utils/bruker_exceptions.py
- Removed duplicate base_reader.py and registry.py from readers/bruker/core/
- Eliminated entire readers/bruker/core/ directory to maintain single core architecture
- Updated all imports across codebase to use consolidated files
- Maintains backward compatibility while simplifying project structure
