--- msiconvert/readers/bruker/core/__init__.py
@@ -1,14 +1,0 @@
-"""Core interfaces and base classes."""
-
-from .base_reader import BaseMSIReader
-from .exceptions import BrukerReaderError, DataError, SDKError
-from .registry import get_reader, register_reader
-
-__all__ = [
-    "BaseMSIReader",
-    "BrukerReaderError",
-    "SDKError",
-    "DataError",
-    "register_reader",
-    "get_reader",
-]
