+++ msiconvert/readers/bruker/sdk/dll_manager.py
@@ -18,7 +18,7 @@ if platform.system() == "Windows":
 else:
     windll = None
 
-from ..core.exceptions import ConfigurationError, SDKError
+from ....utils.bruker_exceptions import ConfigurationError, SDKError
 from .platform_detector import PlatformDetector, get_dll_paths, validate_library_path
 
 logger = logging.getLogger(__name__)
