+++ msiconvert/readers/bruker/sdk/sdk_functions.py
@@ -21,7 +21,7 @@ from typing import List, Optional, Tuple
 
 import numpy as np
 
-from ..core.exceptions import SDKError
+from ....utils.bruker_exceptions import SDKError
 from .dll_manager import DLLManager
 
 logger = logging.getLogger(__name__)
